started reworking the action factory
Going to test a built in validity checker for the function. Also added a big comment explaining how we can use our heuristic function to determine how deep the function should try to search for performance. Such as shallower starting searches and narrowing the search as the branching factor reduces